move message event to generic spot
