+++ src/zelem/index.ts
@@ -1,6 +1,6 @@
 import { RTMClient } from "@slack/rtm-api";
 import { WebClient, WebAPICallResult } from '@slack/web-api';
-import { MessageEvent } from '../message-event';
+import { MessageEvent } from '../_models/message-event';
 import { ZELEM_SLACK_KEY } from "../../development-environment";
 
 const token = process.env['SLACK_KEY'] || ZELEM_SLACK_KEY;
